requirements.txt
fastapi
uvicorn
tensorflow==2.10.0
pillow
numpy
jinja2
aiofiles